--- a/creator-node/Dockerfile
+++ b/creator-node/Dockerfile
@@ -5,8 +5,9 @@ WORKDIR /usr/src/app
 VOLUME "/file_storage"
 ENV storagePath=/file_storage
 
+COPY scripts/healthcheck.sh /
 HEALTHCHECK --interval=5s --timeout=5s --start-period=5m --retries=12 \
-    CMD curl -f http://localhost:${port}/health_check || exit 1
+    CMD /healthcheck.sh
 
 # Add the wait script to the image
 # Script originally from https://github.com/ufoscout/docker-compose-wait/releases/download/2.4.0/wait /usr/bin/wait