--- a/creator-node/scripts/healthcheck.sh
+++ b/creator-node/scripts/healthcheck.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-curl -f http://localhost:${port}/health_check || exit 1