--- a/creator-node/scripts/up.sh
+++ b/creator-node/scripts/up.sh
@@ -20,28 +20,10 @@ fi
 mkdir -p compose/env/tmp/file-storage-${ITERATION}
 . compose/env/tmp/shellEnv${ITERATION}.sh
 
-function return_node_modules() {
-    # if ./node_modules has been created and modified, copy it's content to /tmp
-    if [[ -d node_modules ]]; then
-        mv node_modules/* /tmp/cn-node_modules/ || true
-    fi
-
-    # always ensure no ./node_modules (to prevent ./node_modules/node_modules/*)
-    rm -rf node_modules
-
-    # return node_modules
-    mv /tmp/cn-node_modules node_modules
-}
-
 # build docker image without node_modules
 if [[ "${ITERATION}" == 1 ]]; then
-    # mv ./node_modules away, temporarily
-    mv node_modules /tmp/cn-node_modules
-
     # build image and always return ./node_modules
-    time docker-compose -f compose/docker-compose.yml build \
-        && return_node_modules \
-        || (return_node_modules && exit 1)
+    time docker-compose -f compose/docker-compose.yml build
 fi
 
 . compose/env/tmp/shellEnv${ITERATION}.sh