--- a/service-commands/src/commands/service-commands.json
+++ b/service-commands/src/commands/service-commands.json
@@ -243,18 +243,13 @@
     "port": 4000,
     "health_check_endpoint": "health_check",
     "up": [
-      "cd creator-node; mkdir -p compose/env/tmp/file-storage-#; . compose/env/tmp/shellEnv#.sh; docker-compose -f compose/docker-compose.yml up --build -d",
-      "cd creator-node; . compose/env/unsetShellEnv.sh"
+      "${PROTOCOL_DIR}/creator-node/scripts/up.sh #"
     ],
     "down": [
       "cd creator-node; . compose/env/tmp/shellEnv#.sh; rm -rf compose/env/tmp/file-storage-#; docker-compose -f compose/docker-compose.yml down --remove-orphans"
     ],
     "restart": [
-      "cd creator-node; . compose/env/unsetShellEnv.sh",
-      "cd creator-node; . compose/env/tmp/shellEnv#.sh; docker-compose -f compose/docker-compose.yml down --remove-orphans",
-      "cd libs/; npm run init-local update-cnode-config #",
-      "cd creator-node; . compose/env/tmp/shellEnv#.sh; docker-compose -f compose/docker-compose.yml up --build -d",
-      "cd creator-node; . compose/env/unsetShellEnv.sh"
+      "RESTART=true ${PROTOCOL_DIR}/creator-node/scripts/up.sh #"
     ],
     "register": [
       "cd libs/; npm run init-local register-cnode #"