### Result Summary

`master`:

```
All services brought up in 295.219s
```

`both`:

```
All services brought up in 223.889s
```

For a total of 72 seconds quicker, or 3.7 minutes total.

### Description

Currently, the `Building web server` job is slow and seems to take longer when a remote-dev box has been running for a while.

Additionally, `Building web server` jobs should be SUPER quick when iterating over local changes.

In order to diagnose, test, and modify our build flow, it was FAR easier to understand and manipulate our flow by moving our service commands into a bash script.

Dockerfile changes:
* better utilization of cached layers
* better utilization of multi-stage builds for better cached layer usage
* move no-op comments to `apk add` lines to lock in packages
* remove unneeded `chmod +x`

service-commands.json changes:
* move commands to a bash script for readability
* combine up and restart flows using if-gates
* move `node_modules` away from directory when building to minimize docker's build context

Sister PR for discovery provider:
* #3217


<!--
What is the purpose of this PR?
What is the current behavior? New behavior?
Relevant links and/or information pertaining to PR?
-->


### Tests

I combined and tested with the sister PR off a clean `remote-dev` box:

```
git checkout master
git pull
git checkout -b jc-inf-87-both
git merge jc-inf-87-cn jc-inf-87-dp
```

Both commands came back clean after some ERROR statements during start up:

```bash
d logs cn3_creator-node_1 | grep ERROR
d logs dn1_web-server_1 | grep ERROR | grep -v 'No such file or directory' | grep -v 'could not read redis dump file' | grep -v 'Could not get latest track unavailability'
```

All content nodes now enter the healthy state in dev environments.

```
$ docker ps | grep health
e8f5496e1ca2   audius-identity-service_identity-service_1   Up 31 seconds (healthy)       0.0.0.0:7000->7000/tcp, :::7000->7000/tcp, 0.0.0.0:9229->9229/tcp, :::9229->9229/tcp
fe638c41d9de   cn4_creator-node_1                           Up About a minute (healthy)   0.0.0.0:4003->4003/tcp, :::4003->4003/tcp, 0.0.0.0:9233->9233/tcp, :::9233->9233/tcp
0d9207e8d628   cn3_creator-node_1                           Up About a minute (healthy)   0.0.0.0:4002->4002/tcp, :::4002->4002/tcp, 0.0.0.0:9232->9232/tcp, :::9232->9232/tcp
65f366a7ffad   cn2_creator-node_1                           Up About a minute (healthy)   0.0.0.0:4001->4001/tcp, :::4001->4001/tcp, 0.0.0.0:9231->9231/tcp, :::9231->9231/tcp
897810421b1c   cn1_creator-node_1                           Up 2 minutes (healthy)        0.0.0.0:4000->4000/tcp, :::4000->4000/tcp, 0.0.0.0:9230->9230/tcp, :::9230->9230/tcp
3227bec59ba5   dn1_web-server_1                             Up 3 minutes (healthy)        0.0.0.0:5000->5000/tcp, :::5000->5000/tcp
3598eb0c3e4e   dn1_discovery-provider-db_1                  Up 3 minutes (healthy)        0.0.0.0:5432->5432/tcp, :::5432->5432/tcp
```

### Results

Both tests were performed on a new remote-dev box with a clean `git status`, fresh branches, and `A up && A down` had previously been run to ensure image caches had been built (as they would be with `A --fast`).

Both tests then created dirty repos by running `touch mytest` within the `discovery-provider` and `creator-node` directories to activate Dockerfile's `COPY . .` command to break the image cache. `A up` was then timed:

`master`:

```
All services brought up in 295.219s
```

`both`:

```
All services brought up in 223.889s
```

For a total of 72 seconds quicker.

A second set of tests were done by running `touch mytest2` within the `discovery-provider` and `creator-node` directories to activate Dockerfile's `COPY . .` command to again break the image cache.

```
$ A run discovery-provider restart

discovery-provider 1 - restart | executed in 36.294s
discovery-provider 1 - restart | executed in 26.263s
```

For a total of 10 seconds faster for discovery-provider restarts.

```
$ A run creator-node restart

creator-node 1 - restart | executed in 41.737s
creator-node 1 - restart | executed in 32.149s
```

For a total of 9.6 seconds faster for creator-node restarts.

Trivial time can be saved when we remove `set -x` from `up.sh`. This will be left for a couple of weeks in case we need to debug this release.